+++ NEWS.md
@@ -16,6 +16,8 @@
 
 ## Major Changes
 
+- New function `estimate_counterfactuals()`.
+
 - All `estimate_*()` functions get a `predict` argument, which can be used
   to modulate the type of transformation applied to the predictions (i.e. whether
   predictions should be on the response scale, link scale, etc.). It can also